refactor(sdk): extract abort listener cleanup
